CONTRAST-19386
* Fixed errors caused by merge.
* Moved all UI util methods from Core project to ui project.
* Removed unused interface.
