+++ plugins/com.contrastsecurity.ide.eclipse.ui/src/com/contrastsecurity/ide/eclipse/ui/internal/model/VulnerabilityDetailsPage.java
@@ -45,7 +45,6 @@ import org.eclipse.ui.PlatformUI;
 import com.contrastsecurity.exceptions.UnauthorizedException;
 import com.contrastsecurity.ide.eclipse.core.Constants;
 import com.contrastsecurity.ide.eclipse.core.extended.BaseResponse;
-import com.contrastsecurity.ide.eclipse.core.Util;
 import com.contrastsecurity.ide.eclipse.core.extended.EventResource;
 import com.contrastsecurity.ide.eclipse.core.extended.EventSummaryResource;
 import com.contrastsecurity.ide.eclipse.core.extended.ExtendedContrastSDK;
@@ -96,7 +95,7 @@ public class VulnerabilityDetailsPage extends AbstractPage {
 		GridData gd = new GridData(SWT.FILL, SWT.FILL, true, true);
 		setLayoutData(gd);
 		
-		Util.createLabel(this, ContrastUIActivator.getImage("/icons/contrastLogo.png"));
+		UIElementUtils.createLogoLabel(this, ContrastUIActivator.getImage("/icons/contrastLogo.png"));
 		
 		Composite comboComposite = new Composite(this, SWT.NONE);
 		comboComposite.setLayout(new GridLayout(6, false));
@@ -274,20 +273,6 @@ public class VulnerabilityDetailsPage extends AbstractPage {
 			public void widgetDefaultSelected(SelectionEvent e) {
 			}
 		});
-	}
-
-	private TagsResource getTags(Key key) throws IOException, UnauthorizedException {
-		TagsResource tagsResource = contrastCache.getTagsResources().get(key);
-
-		if (tagsResource == null) {
-			if (key.getTraceId() != null) {
-				tagsResource = getSdk().getTagsByTrace(key.getOrgUuid(), key.getTraceId());
-			} else {
-				tagsResource = getSdk().getTagsByOrg(key.getOrgUuid());
-			}
-			contrastCache.getTagsResources().put(key, tagsResource);
-		}
-		return tagsResource;
 	}
 
 	protected ExtendedContrastSDK getSdk() {
