+++ plugins/com.contrastsecurity.ide.eclipse.ui/src/com/contrastsecurity/ide/eclipse/ui/internal/model/VulnerabilityPage.java
@@ -32,9 +32,9 @@ import org.eclipse.swt.widgets.Listener;
 import com.contrastsecurity.http.ServerFilterForm;
 import com.contrastsecurity.http.TraceFilterForm;
 import com.contrastsecurity.ide.eclipse.core.ContrastCoreActivator;
-import com.contrastsecurity.ide.eclipse.core.Util;
 import com.contrastsecurity.ide.eclipse.ui.ContrastUIActivator;
 import com.contrastsecurity.ide.eclipse.ui.internal.views.VulnerabilitiesView;
+import com.contrastsecurity.ide.eclipse.ui.util.UIElementUtils;
 import com.contrastsecurity.models.Applications;
 import com.contrastsecurity.models.Server;
 import com.contrastsecurity.models.Servers;
@@ -61,7 +61,7 @@ public class VulnerabilityPage extends AbstractPage {
 		GridData gd = new GridData(SWT.FILL, SWT.FILL, true, true);
 		setLayoutData(gd);
 
-		Util.createLabel(this, ContrastUIActivator.getImage("/icons/contrastLogo.png"));
+		UIElementUtils.createLogoLabel(this, ContrastUIActivator.getImage("/icons/contrastLogo.png"));
 
 		Composite comboComposite = new Composite(this, SWT.NONE);
 		comboComposite.setLayout(new GridLayout(5, false));
