--- plugins/com.contrastsecurity.ide.eclipse.ui/src/com/contrastsecurity/ide/eclipse/ui/internal/views/IStatusListener.java
@@ -1,23 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2017 Contrast Security.
- * All rights reserved. 
- * 
- * This program and the accompanying materials are made available under 
- * the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation; either version 3 of the License.
- * 
- * The terms of the GNU GPL version 3 which accompanies this distribution
- * and is available at https://www.gnu.org/licenses/gpl-3.0.en.html
- * 
- * Contributors:
- *     Contrast Security - initial API and implementation
- *******************************************************************************/
-package com.contrastsecurity.ide.eclipse.ui.internal.views;
-
-import com.contrastsecurity.ide.eclipse.core.extended.TraceStatusRequest;
-
-public interface IStatusListener {
-
-	public void onStatusChange(String status, TraceStatusRequest request);
-	
-}
